add: ストア更新処理と諸々追加 - ファイル名更新 - コメント追加
